requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
plotly>=5.19.0
SQLAlchemy>=2.0.27
python-dotenv>=1.0.1
pytest>=8.0.0
black>=24.2.0
flake8>=7.0.0
pymongo>=4.6.1
openpyxl>=3.1.2
reportlab>=4.1.0 